Refactor basic blocks
