+++ src/Stride.Shaders.Spirv/Stride.Shaders.Spirv/Building/Builder.cs
@@ -4,23 +4,10 @@ using Stride.Shaders.Spirv.Core.Buffers;
 
 namespace Stride.Shaders.Spirv.Building;
 
-// Should contain internal data not seen by the client but helpful for the generation like type symbols and other 
-// SPIR-V parameters
-public class SpirvContext(Module module)
-{
-    public int Bound { get; private set; }
-    public Module Module { get; } = module;
-    public SortedList<SymbolType, int> Types { get; } = [];
-}
 
-// Should contain symbols for the SPIR-V module
-public class Module()
-{
-    public SortedList<string, SpirvFunction> Functions { get; init; } = [];
-}
 
 // Should have utility functions to add instruction to the buffer
-public class Builder
+public class Builder()
 {
     public SpirvBuffer Buffer { get; init; } = new();
     public int Position { get; private set; }
@@ -28,11 +15,11 @@ public class Builder
     public void SetPositionTo<TBlock>(TBlock block)
         where TBlock : IInstructionBlock
     {
-        if(block is BasicBlock bb)
+        if (block is BasicBlock bb)
             SetPositionTo(bb.Parent);
         bool blockFound = false;
         Span<int> blockTermination = [
-            (int)SDSLOp.OpBranch,    
+            (int)SDSLOp.OpBranch,
             (int)SDSLOp.OpBranchConditional,
             (int)SDSLOp.OpSwitch,
             (int)SDSLOp.OpReturn,
@@ -45,7 +32,7 @@ public class Builder
         {
             if (e.ResultId is int id && id == block.Id)
                 blockFound = true;
-            if(blockFound && blockTermination.Contains((int)e.OpCode))
+            if (blockFound && blockTermination.Contains((int)e.OpCode))
             {
                 Position = e.WordIndex;
                 return;
