+++ src/Stride.Shaders.Spirv/Stride.Shaders.Spirv/Building/Module.cs
@@ -1,0 +1,8 @@
+namespace Stride.Shaders.Spirv.Building;
+
+
+// Should contain symbols for the SPIR-V module
+public class Module()
+{
+    public SortedList<string, SpirvFunction> Functions { get; init; } = [];
+}
